+++ src/pyedb/configuration/data_model/cfg_nets_data.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class CfgNets:
+    power_ground_nets: list[str] = "default_factory"
+    signal_nets: list[str] = "default_factory"
